# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Ma premi\u00e8re API",
        "description": "Une API permettant l'acc\u00e8s aux donn\u00e9es de la base via une approche RESTful",
        "termsOfService": "",
        "contact": {
            "name": "Guiks",
            "email": "guillaume.delapre@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "myapi.com",
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/personne": {
            "get": {
                "tags": [
                    "personne"
                ],
                "summary": "Get a person list",
                "description": "Use this method to return a listing of persons.",
                "operationId": "listPersonne",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Personne"
                            }
                        }
                    },
                    "204": {
                        "description": "The request didn't return any content."
                    }
                }
            },
            "post": {
                "tags": [
                    "personne"
                ],
                "summary": "Create a person",
                "description": "Use this method to create a new person.",
                "operationId": "createPerson",
                "consumes": [
                    "multipart/form-data",
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "nom",
                        "in": "formData",
                        "description": "Name of the person",
                        "required": true,
                        "type": "string",
                        "maximum": "255"
                    },
                    {
                        "name": "prenom",
                        "in": "formData",
                        "description": "Firstname of the person",
                        "required": true,
                        "type": "string",
                        "maximum": "255"
                    },
                    {
                        "name": "date_naissance",
                        "in": "formData",
                        "description": "Date of birth of the person",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Email of the person",
                        "required": true,
                        "type": "string",
                        "maximum": "255"
                    },
                    {
                        "name": "adresse",
                        "in": "formData",
                        "description": "Address of the person",
                        "required": false,
                        "type": "string",
                        "maximum": "255"
                    },
                    {
                        "name": "cpostal",
                        "in": "formData",
                        "description": "zipcode of the person",
                        "required": true,
                        "type": "string",
                        "maximum": "255"
                    },
                    {
                        "name": "ville",
                        "in": "formData",
                        "description": "city of the person",
                        "required": true,
                        "type": "string",
                        "maximum": "255"
                    },
                    {
                        "name": "pays",
                        "in": "formData",
                        "description": "country of the person",
                        "required": true,
                        "type": "string",
                        "maximum": "255"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Person created",
                        "schema": {
                            "$ref": "#/definitions/Personne"
                        }
                    },
                    "422": {
                        "description": "Champs manquant obligatoire ou incorrect"
                    }
                }
            }
        },
        "/personne/{id_personne}": {
            "get": {
                "tags": [
                    "personne"
                ],
                "summary": "Display a single person",
                "description": "Use this method to return a single person attributes based on its id.",
                "operationId": "showPerson",
                "parameters": [
                    {
                        "name": "id_personne",
                        "in": "path",
                        "description": "id of person to fetch",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/Personne"
                        }
                    },
                    "404": {
                        "description": "Person not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "personne"
                ],
                "summary": "Update a person",
                "description": "Use this method to update the attributes of a person based on its id.",
                "operationId": "updatePerson",
                "consumes": [
                    "multipart/form-data",
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id_personne",
                        "in": "path",
                        "description": "ID person",
                        "required": true,
                        "type": "integer",
                        "maximum": "11"
                    },
                    {
                        "name": "nom",
                        "in": "formData",
                        "description": "Name of the person",
                        "required": false,
                        "type": "string",
                        "maximum": "255"
                    },
                    {
                        "name": "prenom",
                        "in": "formData",
                        "description": "Firstname of the person",
                        "required": false,
                        "type": "string",
                        "maximum": "255"
                    },
                    {
                        "name": "date_naissance",
                        "in": "formData",
                        "description": "Date of birth of the person",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Email of the person",
                        "required": false,
                        "type": "string",
                        "maximum": "255"
                    },
                    {
                        "name": "adresse",
                        "in": "formData",
                        "description": "Address of the person",
                        "required": false,
                        "type": "string",
                        "maximum": "255"
                    },
                    {
                        "name": "cpostal",
                        "in": "formData",
                        "description": "zipcode of the person",
                        "required": false,
                        "type": "string",
                        "maximum": "255"
                    },
                    {
                        "name": "ville",
                        "in": "formData",
                        "description": "city of the person",
                        "required": false,
                        "type": "string",
                        "maximum": "255"
                    },
                    {
                        "name": "pays",
                        "in": "formData",
                        "description": "country of the person",
                        "required": false,
                        "type": "string",
                        "maximum": "255"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Person updated",
                        "schema": {
                            "$ref": "#/definitions/Personne"
                        }
                    },
                    "422": {
                        "description": "Champs manquants obligatoires ou incorrects"
                    }
                }
            },
            "delete": {
                "tags": [
                    "personne"
                ],
                "summary": "Delete a Person",
                "description": "Use this method to delete a person based on its id.",
                "operationId": "deletePerson",
                "parameters": [
                    {
                        "name": "id_personne",
                        "in": "path",
                        "description": "Person ID to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Person deleted"
                    },
                    "404": {
                        "description": "Invalid person value"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Personne": {
            "required": [
                "nom",
                "prenom",
                "date_naissance",
                "email",
                "cpostal",
                "ville",
                "pays"
            ],
            "properties": {
                "id_personne": {
                    "type": "number",
                    "format": "int64",
                    "default": 1
                },
                "nom": {
                    "type": "string",
                    "format": "string",
                    "default": "SOLEVIC"
                },
                "prenom": {
                    "type": "string",
                    "format": "string",
                    "default": "alexis"
                },
                "date_naissance": {
                    "type": "date",
                    "format": "date",
                    "default": "1997-01-05"
                },
                "email": {
                    "type": "string",
                    "format": "string",
                    "default": "solevic.alexis@lycos.fr"
                },
                "adresse": {
                    "type": "string",
                    "format": "string",
                    "default": "1 rue Ketanou"
                },
                "cpostal": {
                    "type": "string",
                    "format": "string",
                    "default": "94170"
                },
                "ville": {
                    "type": "string",
                    "format": "string",
                    "default": "LE PERREUX SUR MARNE"
                },
                "pays": {
                    "type": "string",
                    "format": "string",
                    "default": "France"
                }
            },
            "xml": {
                "name": "Personne"
            }
        },
...
    }
}